move ConsoleLogger to separate import file and cleanup
