+++ util/gen_effect_id.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 
+import { ConsoleLogger, LogLevelKey } from './console_logger';
 import { cleanName } from './csv_util';
-import { ConsoleLogger, LogLevelKey } from './generate_data_files';
 import { OutputFileAttributes, XivApi } from './xivapi';
 
 const _EFFECT_ID: OutputFileAttributes = {
@@ -36,7 +36,8 @@ type OutputEffectId = {
 };
 
 const _SCRIPT_NAME = path.basename(import.meta.url);
-let log: ConsoleLogger;
+const log = new ConsoleLogger();
+log.setLogLevel('alert');
 
 // TODO: add renaming?
 // Almagest: 563
@@ -162,7 +163,6 @@ const assembleData = (apiData: XivApiStatus): OutputEffectId => {
 };
 
 export default async (logLevel: LogLevelKey): Promise<void> => {
-  log = new ConsoleLogger();
   log.setLogLevel(logLevel);
   log.info(`Starting processing for ${_SCRIPT_NAME}`);
